moved x2sa logic functions to it's own service
