+++ src/app/services/x2sa/x2sa.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { X2saService } from './x2sa.service';
+
+describe('X2saService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: X2saService = TestBed.get(X2saService);
+    expect(service).toBeTruthy();
+  });
+});
