+++ src/app/tab1/tab1.page.html
@@ -34,5 +34,5 @@
 
   <ion-button size="large" routerLink="/settings">Einstellungen</ion-button>
 
-  <ion-button (click)="connectToBank()">XS2A Test mit Sandbox</ion-button>
+  <ion-button (click)="x2saService.connectToBank()">XS2A Test mit Sandbox</ion-button>
 </ion-content>
